# Branch-after-load scheme for OCaml memory model on Arm64