@gasche 2024-08-24
I wonder about the performance impact of the more-expensive OCaml-C boundary. If there is an easy way to run Sandmark benchmarks on a arm64 machine, it would be interesting to look at the results.