@xavierleroy 2024-08-25
I cannot accept that we're making loads from immutable data structures more costly.  This is shifting the costs of shared-memory parallelism from where it belongs (concurrent mutation of data structures) to where it does not belong (reading immutable data).  Please don't dump costs on purely functional programming.