@gasche 2024-08-25
If the branch after load is for ordering with future stores to the same location, we may be able to avoid it for immutable locations?

This is probably harder than it sounds because I guess there are situations where we mutate immutable data types, for example in recursive backpatching schemes. But I guess we could use fence-before-store for immutable locations and branch-after-load for mutable locations -- basically a cheap read barrier on mutable locations.