@lthls 2024-08-25
> If the branch after load is for ordering with future stores to the same location, we may be able to avoid it for immutable locations?

We don't know which reads are mutable or immutable in the backend, with a few exceptions like loads from closures.
I don't remember why we discard this information (in Cmmgen, if I remember correctly), but I suspect that the backend makes assumptions on immutable loads that are not true in general for regular immutable values.