@gasche 2024-08-25
We don't need to have the exact same notion of immutability as the optimizer needs, only to propagate information that assigns one or the other scheme to each memory operation, and the correctness requirement is that we don't "race" by using the two schemes on any given location, without a barrier between them. (In particular, if we use `Obj.magic` to freeze a mutable value into an immutable value, we would need a barrier at freezing point, or at least between the last mutable store and the first immutable load.)